[FIX] website_event_track_online: miss-assigned var. in ServiceWorker
When caching a POST request a cache key is builded based on a serialized
version of the request... but the serialization wasn't properly
performed.

This commit properly assigns the `serializedRequest` variable before
attempting to build the cache key.

This was due to a miss-extraction of that variable (see commit
odoo/odoo@9a570d034603958d593135ddde7f49c3f9944b66 ).

closes odoo/odoo#56049

X-original-commit: f5c0f076a6213950b507b00e28c887746c10ae59
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>