[IMP] website_event_track_online: PWA adds prefetch for offline use
This commit adds the prefetch of the current page's children (cf. under
the same scope) to be available for offline use.

For performance reason, only the first-level children are added to the
cache to avoid a huge amount of requests and/or resources used by simply
loading the event website's pages. Also, for the same reason, this
process is delegated to the ServiceWorker to avoid cluttering the Main
Thread.

Note: due to some security restriction introduced in the Fetch spec,
some redirects may prevent offline-requests from being properly
fulfilled from the cache. See references for further details.

References:
- https://github.com/whatwg/fetch/issues/763
- https://github.com/GoogleChromeLabs/sw-precache/issues/220
- https://fetch.spec.whatwg.org/#concept-request-redirect-mode

closes odoo/odoo#56159

X-original-commit: 8b06f6bf89e29d7d5b4e64defbbc0b655778db00
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>