[FIX] hr_timesheet: fallback on employee in report
The report "Timesheet Entries" is added in hr_timesheet.

employee_id is added on analytic line (used for timesheet lines) in
hr_timesheet also, and on timesheets view, we display employee.

But in the report, we display "user_id.partner_id.name" for the
responsible, which only works for employee with users.

In this commit we fallback on the employee name if he doesn't have an
associated res.user (for stability keeping original value by default).

opw-2322050
closes #56315

closes odoo/odoo#56330

X-original-commit: c6b8d792f410d049a5a4566718ae1c9bffef7ec0
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>