[FIX] base: Render "custom colors" widget
Since commit 72950552ba5cc486849202b1dfd1f6c4943d3a78 the invisible form
widgets are not rendered at all. This caused an issue in the document
layout designer where the "custom_colors" field was invisible but still
referenced by a label with a "for" attribute.

Now, said field is not completely invisible anymore; it is set with
"display:none" instead to be rendered invisible and interactive as
before.

closes odoo/odoo#55970

Task: 2320982
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>