[FIX] mail: ignore broadcast notifications targeting another user
This is typically not the case in production code, but it doesn't hurt to be
safe. The main reason for the change however is that it causes issues in tests
due to the easiest way to simulate notifications being to send them directly on
the bus which is "listened to" no matter the target user.

PR: #56243
Part of task-2323670

X-original-commit: 1c1f76c7538952f1b7949e6a56d788f8ce769d77