[FIX] account: avoid long recomputation recursion
when creating the first bank statement on a journal. It will trigger
manual field recomputations:
https://github.com/odoo/odoo/blob/9444e36b8fe5ec3ed07ddbbbcb76ad7dc4091827/addons/account/models/account_bank_statement.py#L288-L297

previous_statement_id will correctly be False
The manual recomputation in create will recalculate all bank
statements that have a False previous_statement_id.
These bank statements are usually the first ones in a journal. When they get
recomputed it triggers recomputations for every bank statement that
follows it.

This commit try to limit this behavior by recomputing statements only in
the same journal

opw-2311173

closes odoo/odoo#56648

X-original-commit: e011276c2dcf82794b787580d46732262208e8c3
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>
Signed-off-by: agr-odoo <agr-odoo@users.noreply.github.com>