[FIX] web: preview PDF file on iOS
Before this commit, the end users were unable to click on the
kanban image to see document's preview.

To fix this we removed lines from a previous fix which is no
longer useful in 13.0 because we now use kanban views for
many2X fields instead of list views.
See original commit:
https://github.com/odoo/odoo/commit/e3ac6ca2251f095895871b6f3b86406ee3525f40
So for mobile phones it's working fine.

For iPads, this rule is no longuer necessary since iPadOS.
See "Accelerated Scrolling on iOS and iPadOS":
https://webkit.org/blog/9674/new-webkit-features-in-safari-13/
For previous versions it still works because the DOM has changed.

Steps to reproduce:
- Go to Document
- Click on "o_kanban_image"
=> Your're stuck on a grayed page.

opw-2260101

closes odoo/odoo#56977

X-original-commit: 3c6b9b291aa1f4695cc466646764fe7cb41f1d65
Related: odoo/enterprise#12907
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>