[IMP] account_edi: improve test helper
Allowing to filter the type of EDI attachments data we want to filter allows finer testing when multiple EDI modules are installed. Wihtout that, we risk to also consider data from other formats than the one we are testing (it actually happened for l10n_mx_edi).

closes odoo/odoo#56824

X-original-commit: 719ffecb58a7e81089427c9e22b5e5f99246980b
Related: odoo/enterprise#12811
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>