[FIX] mail: allow to create next activities when not having a clear model as context
Before this commit, when we tried to launch next activity from the already
scheduled actvity's popover on chatter (activity that do have a next activity
set on it), it resulted into a traceback. This was due to a recent commit[1]
trying to set the default activity type.

This commit fixes the issue by improving method setting the default activity
type and thus allowing to schedule next activity from popover.

[1] - https://github.com/odoo/odoo/commit/47063ceca3e221a3045de84fea66d3664607d207

Task ID-2302572

X-original-commit: 0cd3c5475d41579f67903317b405430a4affa885